Update email.test.php
Test 'testEmailSend' stubs two methods in the Email class  ("logMail","phpMail").

I stubbed the phpMail meethod because my test environment did not include an email server, so the call to Mail() was failing.  In order to make the stub possible, I changed the phpMail() method definition in email.php from "private" to "protected".